@@ -1,16 +1,18 @@
 name: pub_semver
-version: 1.4.5-dev
-
+version: 2.0.0-nullsafety.0
 description: >-
  Versions and version constraints implementing pub's versioning policy. This
  is very similar to vanilla semver, with a few corner cases.
-homepage: https://github.com/dart-lang/pub_semver
+repository: https://github.com/dart-lang/pub_semver
 
 environment:
- sdk: '>=2.0.0 <3.0.0'
+ sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
-  collection: ^1.0.0
+  collection: ^1.15.0-nullsafety.2
 
 dev_dependencies:
-  test: ^1.0.0
+  test: ^1.16.0-nullsafety.1
+
+dependency_overrides:
+  analyzer: any
